refactored tests and moved decoders to the decoder file. Decoder tested successfully, 40/40 passing. TX/RX end-to-end not tested. Hotfix for cdh to remove decimal place on the psutil percentages to align with encoder.
