+++ src/utils/tests/test_subsystems/test_cdh.py
@@ -1,23 +1,7 @@
-import pytest
-import struct
 from src.subsystems import cdh
-from src.ccsds import encoder
+from src.ccsds import encoder, decoder
 
 # Temporary function to decode the payload
-def decode_ccsds_cdh_payload(payload: bytes) -> dict:
-    fields = struct.unpack(">fffBBfIHBB", payload)
-    return {
-        "processor_temp": fields[0],
-        "processor_freq": fields[1],
-        "processor_util": fields[2],
-        "ram_usage": fields[3],
-        "disk_usage": fields[4],
-        "cooling_fan_speed": fields[5],
-        "uptime": fields[6],
-        "watchdog_counter": fields[7],
-        "software_version": fields[8],
-        "event_flags": fields[9],
-    }
 
 # Test cases for the CDH subsystem telemetry
 def test_cdh_telemetry_keys():
@@ -46,7 +30,7 @@ def test_cdh_payload_encoding_decoding():
     """
     data = cdh.get_cdh_telemetry()
     payload = encoder.encode_ccsds_cdh_payload(data)
-    decoded = decode_ccsds_cdh_payload(payload)
+    decoded = decoder.decode_ccsds_cdh_payload(payload)
 
     # approximate check, float rounding
     assert abs(decoded['processor_temp'] - data['processor_temp']) < 0.5
