+++ src/utils/tests/test_subsystems/test_propulsion.py
@@ -1,26 +1,5 @@
 from src.subsystems import propulsion
-from src.ccsds import encoder
-import struct
-
-def decode_ccsds_propulsion_payload(payload: bytes) -> dict:
-    """
-    Decode the propulsion payload to verify correctness.
-    """
-    fields = struct.unpack(">ffffBBBBffBB", payload)
-    return {
-        "fuel_level": fields[0],
-        "oxidizer_level": fields[1],
-        "tank_pressure": fields[2],
-        "feedline_temp": fields[3],
-        "valve_status": fields[4],
-        "thruster_firing": fields[5],
-        "thruster_mode": fields[6],
-        "propulsion_fault_flags": fields[7],
-        "rcs_tank_level": fields[8],
-        "rcs_tank_pressure": fields[9],
-        "rcs_thruster_status": fields[10],
-        "rcs_fault_flags": fields[11]
-    }
+from src.ccsds import encoder, decoder
 
 def test_propulsion_telemetry_keys():
     """
@@ -50,7 +29,7 @@ def test_propulsion_payload_encoding_decoding():
     """
     data = propulsion.get_propulsion_telemetry()
     payload = encoder.encode_ccsds_propulsion_payload(data)
-    decoded = decode_ccsds_propulsion_payload(payload)
+    decoded = decoder.decode_ccsds_propulsion_payload(payload)
 
     # floats, allow small differences
     for field in [
